refactor: 상호 관련된 컴포넌트, 훅, 유틸을 응집도 있게 배치
