+++ src/_components/Home/SubscribeModal/_utils/validateEmail.ts
