+++ src/_components/common/Paginator/_hooks/useDeviceWidth.ts
